+++ src/instrumentman/filetransfer/__init__.py
@@ -23,15 +23,15 @@ from ..utils import (
 @com_port_argument()
 @argument(
     "directory",
-    help="directory to list files in (path should end with '/')",
+    help="Directory to list files in",
     type=str,
-    default="/"
+    default=""
 )
 @com_option_group()
 @option(
     "-d",
     "--device",
-    help="memory device",
+    help="Memory device",
     type=Choice(
         (
             "internal",
@@ -47,7 +47,7 @@ from ..utils import (
 @option(
     "-f",
     "--filetype",
-    help="file type (ignored in recursive mode)",
+    help="File type (all files are shown when not set)",
     type=Choice(
         (
             "image",
@@ -65,14 +65,26 @@ from ..utils import (
 @option(
     "--depth",
     help=(
-        "recursive depth "
+        "Recursive depth "
         "(0: unlimited; 1<=x: depth of directory search)"
     ),
     type=IntRange(0),
     default=1
 )
 def cli_list(**kwargs: Any) -> None:
-    """List files on an instrument."""
+    """
+    List files on an instrument.
+
+    For each file in the specified directory (and discovered subdirectories
+    when recursive search is enabled) the file name, file size and the time of
+    last modification is displayed. Not empty directories and items, that are
+    likely directories are show in blue and light blue colors. Known text
+    formats are shown green. Image and drawing formats are shown in magenta.
+    Database files are red. Other files are shown without special color.
+
+    This command requires a GeoCom capable instrument, that supports file
+    operations (TPS1200 and later).
+    """
     from .app import main_list
 
     main_list(**kwargs)
@@ -87,21 +99,21 @@ def cli_list(**kwargs: Any) -> None:
 @argument(
     "filename",
     help=(
-        "file to download (including path with '/' separators if filetype "
+        "File to download (including path with '/' separators if filetype "
         "option is not specified)"
     ),
     type=str
 )
 @argument(
     "output",
-    help="file to save downloaded data to",
+    help="File to save downloaded data to",
     type=File("wb", lazy=False)
 )
 @com_option_group()
 @option(
     "-d",
     "--device",
-    help="memory device",
+    help="Memory device",
     type=Choice(
         (
             "internal",
@@ -117,7 +129,7 @@ def cli_list(**kwargs: Any) -> None:
 @option(
     "-f",
     "--filetype",
-    help="file type",
+    help="File type (full file path is required if this option is not set)",
     type=Choice(
         (
             "image",
@@ -137,17 +149,27 @@ def cli_list(**kwargs: Any) -> None:
 @option(
     "-c",
     "--chunk",
-    help="chunk size (max 450 for normal and 1800 for VivaTPS large download)",
-    type=IntRange(1, 1800),
-    default=450
+    help="Chunk size (max. 225 for normal and 900 for large download mode)",
+    type=IntRange(1, 900),
+    default=225
 )
 @option(
     "--large",
-    help="use large download commands (only available from VivaTPS)",
+    help="Use large download commands (only available from VivaTPS)",
     is_flag=True
 )
 def cli_download(**kwargs: Any) -> None:
-    """Download a file from the instrument."""
+    """
+    Download a file from the instrument.
+
+    Any format can be transferred. The file is downloaded in chunks of hex
+    encoded binary data. The speed is strongly dependent on the connection
+    baud and chunk size. Use the highest baud supported by the instrument, and
+    the largest chunk size for the fastest download.
+
+    This command requires a GeoCom capable instrument, that supports file
+    operations (TPS1200 and later).
+    """
     from .app import main_download
 
     main_download(**kwargs)